Major refactor.
Removed 'src' directory from the cdk project. The 'bdd' folder is where all the logic will be.

Refactored the cdk project to use 'workflows' as the primary method of programming lambda functions. The service objects are separate from the workflows. The lambda function handlers are the workflows and the service objects are used to carry out the individual steps to complete the workflow.